move TSObjects to separate file
